+++ resources/js/Pages/Shop/Subscription.vue
@@ -209,7 +209,7 @@ async function submit() {
     })
 
     //Submit the form
-    form.post('/shop/subscribe')
+    form.post('/contribute/subscription')
 
   }
 }